回复问题or评论异常

+++ src/main/java/com/hciot/community/controller/PublishController.java
@@ -1,11 +1,9 @@
 package com.hciot.community.controller;
 
 import com.hciot.community.dto.QuestionDTO;
-import com.hciot.community.mapper.QuestionMapper;
 import com.hciot.community.model.Question;
 import com.hciot.community.model.User;
 import com.hciot.community.service.QuestionService;
-import com.sun.org.apache.xpath.internal.operations.Mod;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -23,7 +21,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                        Model model){
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title",question.getTitle());
@@ -43,7 +41,7 @@ public class PublishController {
             @RequestParam(value = "title", required = false) String title,
             @RequestParam(value = "description", required = false) String description,
             @RequestParam(value = "tag", required = false) String tag,
-            @RequestParam(value = "id", required = false) Integer id,
+            @RequestParam(value = "id", required = false) Long id,
             HttpServletRequest request,
             Model model){
         model.addAttribute("title",title);
